+++ KiranaStoreUI/Models/Stock.cs
@@ -31,11 +31,11 @@ namespace KiranaStoreUI.Models
 
         [Required(ErrorMessage = "Quantity is required")]
         [Range(0, int.MaxValue, ErrorMessage = "Quantity cannot be negative")]
-        public int Quantity { get; set; }  // Current stock count
+        public decimal Quantity { get; set; }  // Current stock count
 
         [Required(ErrorMessage = "Minimum Quantity is required")]
         [Range(0, int.MaxValue, ErrorMessage = "Minimum Quantity cannot be negative")]
-        public int MinimumQuantity { get; set; }  // For low stock alert
+        public decimal MinimumQuantity { get; set; }  // For low stock alert
 
         // Navigation property
         public Product Product { get; set; }
